Create OZ governor wizard and token section
